+++ test/ocaml/NonGenericType.mli
@@ -1,0 +1,8 @@
+type nonGenericType =
+  { ngA : string
+  ; ngB : int
+  }
+
+val encodeNonGenericType : nonGenericType -> Js_json.t
+
+val decodeNonGenericType : Js_json.t -> (nonGenericType, string) Js_result.t